business logic, some temp exceptions, trying to get compose to work properly

+++ src/main/java/com/example/todo_app/TaskService.java
@@ -28,19 +28,39 @@ public class TaskService {
         return taskRepository.findById(id);
     }
 
-    public TaskDto addTask(TaskDto taskDto) {
-        taskDto.setId(UUID.randomUUID()); //todo in db generation
+    public TaskDto addTask(TaskDto taskDto) throws Exception {
+        check10();
+        taskDto.setId(UUID.randomUUID()); //todo in db generation?
 
         return taskRepository.save(taskDto);
     }
 
-    public TaskDto updateTask(TaskDto taskDto) {
-        //todo check exists
-
-        return taskRepository.save(taskDto);
+    //check if 10 incomplete tasks already exist
+    private void check10() throws Exception {
+        List<TaskDto> byCompletedFalse = taskRepository.findByCompletedFalse();
+        if (byCompletedFalse.size() >= 10) {
+            throw new Exception("placeholder 10 exception");
+        }
     }
 
-    public void deleteTask(UUID taskId) {
+    public TaskDto updateTask(TaskDto newTask) throws Exception {
+        Optional<TaskDto> oldTask = taskRepository.findById(newTask.getId());
+        if (oldTask.isPresent()) {
+            if (oldTask.get().isCompleted() && !newTask.isCompleted()) {
+                check10();
+            }
+        } else {
+            throw new Exception("placeholder missing task to update/delete");
+        }
+
+        return taskRepository.save(newTask);
+    }
+
+    public void deleteTask(UUID taskId) throws Exception {
+        Optional<TaskDto> oldTask = taskRepository.findById(taskId);
+        if (oldTask.isEmpty()) {
+            throw new Exception("placeholder missing task to update/delete");
+        }
         taskRepository.deleteById(taskId);
     }
